frog-fifo-v1: distill progress guarantees in set_barrier
It is not clear what it means for a surface to be presented. For
example, consider a visible surface that is partially on screen and
partially off screen. It is not unreasonable to say that such a surface
is being presented. But if the content update only damages a region that
is off screen, then the latching deadline is unlikely to occur soon and
the surface might in fact dead lock itself.

Instead of trying to define what _being presented_ means in this
context, reword to get closer to the goal of the sentence.

In practice, on a regular monitor (VRR or otherwise), compositors are
more likely to schedule a latching deadline unconditionally if a content
update contains a set_barrier request. But the protocol does not need to
concern itself with this.

Signed-off-by: Julian Orth <ju.orth@gmail.com>